chore: 小修小补 (#2905)
* refactor(launch): simplify ram allocation flow

* refactor(controls): remove scroll viewer goto

* fix(download): correct optifabric warning condition

* refactor(ui): use short-circuit boolean operators

* refactor(mods): use string length checks

* refactor: use short-circuit logical operators

* fix(controls): continue scroll viewer search

* refactor(goto): remove simple control flow labels

+++ Plain Craft Launcher 2/Pages/PageInstance/PageInstanceInstall.xaml.cs
@@ -145,13 +145,13 @@ public partial class PageInstanceInstall
         if (!ModDownloadLib.McInstall(Request, BtnSelectStart.Text.AfterFirst("开始")))
             return;
         // 删除旧的独立 API 文件
-        if (SelectedFabricApi is not null & _currentFabricApiPath is not null)
+        if (SelectedFabricApi is not null && _currentFabricApiPath is not null)
             File.Delete(_currentFabricApiPath);
-        if (SelectedLegacyFabricApi is not null & _currentLegacyFabricApiPath is not null)
+        if (SelectedLegacyFabricApi is not null && _currentLegacyFabricApiPath is not null)
             File.Delete(_currentLegacyFabricApiPath);
-        if (SelectedQSL is not null & _currentQslPath is not null)
+        if (SelectedQSL is not null && _currentQslPath is not null)
             File.Delete(_currentQslPath);
-        if (SelectedOptiFabric is not null & _currentOptiFabricPath is not null)
+        if (SelectedOptiFabric is not null && _currentOptiFabricPath is not null)
             File.Delete(_currentOptiFabricPath);
         // 返回主页
         ModMain.FrmMain.PageChange(new FormMain.PageStackData { Page = FormMain.PageType.Launch });
@@ -857,7 +857,7 @@ public partial class PageInstanceInstall
         else
             HintQuiltFabricAPI.Visibility = Visibility.Collapsed;
 
-        if (SelectedFabric is not null | SelectedLegacyFabric is not null && SelectedOptiFine is not null &&
+        if ((SelectedFabric is not null || SelectedLegacyFabric is not null) && SelectedOptiFine is not null &&
             SelectedOptiFabric is null)
         {
             if (VanillaDrop >= 140 && VanillaDrop <= 150)
